Rename rules to match eslint conventions
