+++ lib/rules/destructure-ember-namespace.js
