+++ lib/rules/no-function-prototype-extension-calls.js
